requirements.txt
matplotlib>=3.5.1
numpy>=1.22.3
Pillow>=9.1.0
torch>=1.11.0
torchvision>=0.15.2
gpytorch>=1.6.0
botorch>=0.2.1
pandas>=2.0.3
mario-gpt>=0.1.3
cma>=3.3.0
seaborn>=0.12.2
tqdm>=4.65.0
wheel>=0.38.4
transformers>=4.31.0
